requirements.txt
beautifulsoup4==4.11.1
emoji==2.8.0
matplotlib==3.5.3
pandas==1.4.4
Pillow==9.2.0
plotly==5.14.1
seaborn==0.12.1
spacy==3.4.2
tqdm==4.64.0
wordcloud==1.9.2
